requirements.txt
ultralytics==8.0.196
opencv-python-headless==4.8.1.78
easyocr==1.7.0
numpy==1.24.3
Pillow==10.0.1
